Move provider classes to maven surefire package
