+++ surefire-providers/surefire-junit5/src/test/java/org/apache/maven/surefire/junit5/RunListenerAdapterTests.java
@@ -1,4 +1,4 @@
-package org.junit.platform.surefire.provider;
+package org.apache.maven.surefire.junit5;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -29,7 +29,6 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.util.Collections;
 import java.util.Optional;
 
 import org.apache.maven.surefire.report.ReportEntry;
